Refactored the cli commands.
This creates a new 'configure' command that is used to generate
configurations like a docker-compose.yml file.  This allows 'build' to
be used for building the service images.
