+++ src/gantry/cli/__init__.py
@@ -1,0 +1,4 @@
+from ._main import main
+
+if __name__ == '__main__':
+    main()
